+++ docs/changelog.md
@@ -1,2 +1,5 @@
+# TBD
+* Add build infra
+
 # 0.1.0
 * Initial commit